+++ OfficeHVAC.Messages/RoomStatusMessage.cs
@@ -4,9 +4,9 @@
     {
         public string RoomName { get; private set; }
 
-        public float Temperature { get; private set; }
+        public double Temperature { get; private set; }
 
-        public RoomStatusMessage(string roomName, float temperature)
+        public RoomStatusMessage(string roomName, double temperature)
         {
             RoomName = roomName;
             Temperature = temperature;
